requirements.txt
discord.py == 1.3.2
PyNaCl == 1.3.0
asyncio == 3.4.3
urllib3 == 1.25.8
cogs == 